Added more 7zip wrapper and callbacks

+++ kxf/Compression/Common.h
@@ -5,7 +5,7 @@
 
 namespace kxf::Compression
 {
-	using FileIndex = uint32_t;
+	using FileIndex = size_t;
 	using FileIndexVector = std::vector<FileIndex>;
 
 	constexpr FileIndex InvalidFileIndex = std::numeric_limits<FileIndex>::max();
